[Model Runner V2] Add sample/ directory and reorganize files
